added xml file, remade main-original file
